pyproject.toml
[project]
name = "raian-bot"
version = "0.15.0"
description = "simple qq bot bases on Araidne and Alconna"
authors = [
    {name = "RF-Tar-Railt", email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "arclet-alconna<2.0.0, >=1.3.2",
    "arclet-alconna-graia<1.0.0, >=0.9.0",
    "arclet-alconna-tools>=0.2.1",
    "graia-ariadne<0.10.0, >=0.9.4",
    "graia-amnesia>=0.6.0",
    "graia-broadcast>=0.18.2",
    "graia-saya>=0.0.16",
    "graia-scheduler>=0.0.8",
    "unwind~=0.4.0",
    "Pillow>=9.0.1",
    "aiohttp~=3.8.1",
    "pyquery~=1.4.3",
    "pydantic~=1.9.0",
    "picimagesearch>=3.7.0",
    "PyYAML~=6.0",
    "loguru~=0.6.0",
    "psutil~=5.9.0",
    "tencentcloud-sdk-python==3.0.495",
    "ujson~=5.1.0",
    "creart~=0.2.1",
    "graiax-playwright>=0.2.0.post1",
    "graiax-silkcoder[ffmpeg]>=0.3.4",
    "graiax-text2img-playwright>=0.1.0b5",
    "diro-py>=0.1.1",
    "httpx>=0.23.0",
    "arknights-toolkit>=0.3.0",
    "graiax-fastapi>=0.2.0",
    "uvicorn>=0.18.3",
    "graiax-shortcut>=0.2.0"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "AGPL-3.0"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]

[tool.black]
include = '\.pyi?$'
line-length = 120
target-version = ['py38']